requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.7
djangorestframework==3.11.2
flake8==3.8.4
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytz==2021.1
PyYAML==5.3.1
sqlparse==0.4.1
toml==0.10.2
